refactor: option の命名変更と api のディレクトリの移動
